Move command debugging

+++ Core/Inc/lre_stepper.h
@@ -4,4 +4,4 @@ void lre_stepper_setStep(uint8_t step);
 
 void lre_stepper_setStep_side(uint8_t step, uint8_t right);
 
-void rotate(uint8_t step);
+void rotate(uint32_t step);
